requirements.txt
pyinotify
SpeechRecognition
